% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{adae}
\alias{adae}
\title{ADAE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 848 rows and 50 columns.
}
\source{
<https://github.com/Hbrdata/BioStatsSuite/blob/master/data-raw/adae.sas7bdat>
}
\usage{
adae
}
\description{
ADAE
}
\keyword{datasets}
