% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{adcrslb}
\alias{adcrslb}
\title{ADCRSLB}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 39600 rows and 39 columns.
}
\source{
<https://github.com/Hbrdata/BioStatsSuite/blob/master/data-raw/adcrslb.xlsx>
}
\usage{
adcrslb
}
\description{
ADCRSLB
}
\keyword{datasets}
