% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{ADSL}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 220 rows and 11 columns.
}
\source{
<https://github.com/Hbrdata/BioStatsSuite/blob/master/data-raw/adsl.xlsx>
}
\usage{
adsl
}
\description{
ADSL
}
\keyword{datasets}
