% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{adts}
\alias{adts}
\title{ADTS}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34320 rows and 56 columns.
}
\source{
<https://github.com/Hbrdata/BioStatsSuite/blob/master/data-raw/adts.sas7bdat>
}
\usage{
adts
}
\description{
ADTS
}
\keyword{datasets}
