% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{cov_adur}
\alias{cov_adur}
\title{COV_ADUR}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 480 rows and 30 columns.
}
\source{
<https://github.com/Hbrdata/BioStatsSuite/blob/master/data-raw/cov_adur.xlsx>
}
\usage{
cov_adur
}
\description{
COV_ADUR
}
\keyword{datasets}
