% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{ex2_vist}
\alias{ex2_vist}
\title{EX2_VIST}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 720 rows and 19 columns.
}
\source{
<https://github.com/Hbrdata/BioStatsSuite/blob/master/data-raw/ex2_vist.sas7bdat>
}
\usage{
ex2_vist
}
\description{
EX2_VIST
}
\keyword{datasets}
