% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_reader.R
\name{get_data_name}
\alias{get_data_name}
\title{Extract data name from file name}
\usage{
get_data_name(file_name)
}
\arguments{
\item{file_name}{File name}
}
\value{
Base name without extension
}
\description{
Extract data name from file name
}
