% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_reader.R
\name{get_default_data_name}
\alias{get_default_data_name}
\title{Get default data name based on analysis type}
\usage{
get_default_data_name(analysis_type)
}
\arguments{
\item{analysis_type}{Type of analysis}
}
\value{
Default data name
}
\description{
Get default data name based on analysis type
}
