% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_reader.R
\name{get_example_data_path}
\alias{get_example_data_path}
\title{Get example data file path based on analysis type}
\usage{
get_example_data_path(analysis_type)
}
\arguments{
\item{analysis_type}{Type of analysis}
}
\value{
Path to the example data file
}
\description{
Get example data file path based on analysis type
}
