% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_reader.R
\name{get_file_type}
\alias{get_file_type}
\title{Get data file type}
\usage{
get_file_type(file_name)
}
\arguments{
\item{file_name}{File name}
}
\value{
File extension in lowercase
}
\description{
Get data file type
}
