% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_reader.R
\name{read_data_file}
\alias{read_data_file}
\title{Read data file based on file extension}
\usage{
read_data_file(
  file_path,
  file_name,
  csv_separator = ",",
  csv_decimal = ".",
  csv_header = TRUE
)
}
\arguments{
\item{file_path}{Path to the file}

\item{file_name}{Name of the file}

\item{csv_separator}{CSV separator character}

\item{csv_decimal}{CSV decimal character}

\item{csv_header}{Whether CSV has header}
}
\value{
Data frame containing the read data
}
\description{
Read data file based on file extension
}
