% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{tmax_demo}
\alias{tmax_demo}
\title{TMAX_DEMO}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 4 columns.
}
\source{
<https://github.com/Hbrdata/BioStatsSuite/blob/master/data-raw/tmax_demo.sas7bdat>
}
\usage{
tmax_demo
}
\description{
TMAX_DEMO
}
\keyword{datasets}
