% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{tyypspa}
\alias{tyypspa}
\title{TYYPSPA}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7920 rows and 32 columns.
}
\source{
<https://github.com/Hbrdata/BioStatsSuite/blob/master/data-raw/tyypspa.sas7bdat>
}
\usage{
tyypspa
}
\description{
TYYPSPA
}
\keyword{datasets}
