% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{adhj}
\alias{adhj}
\title{ADHJ}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 179 rows and 36 columns.
}
\source{
<https://github.com/Hbrdata/BioStatsSuite/blob/master/data-raw/adhj.xlsx>
}
\usage{
adhj
}
\description{
ADHJ
}
\keyword{datasets}
